+++ public/fontes/gamma.json
