+++ public/img/bomba.png
