+++ public/img/diamond.png
