+++ public/telaInicial.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Document</title>
+    <script type="text/javascript" src="phaser.min.js"></script>
+    <script type="text/javascript" src="telaInicial.js"></script>
+    <script type="text/javascript" src="jogo.js"></script>
+    <script type="text/javascript" src="game.js"></script>
+    
+</head>
+    <body>
+
+    </body>
+</html>
